added data tapping and new aggregation function for spatiotemporal data, separated pipeline, config, schema files from other data types
