+++ config/SEspatioTemporalSchema.json
@@ -30,6 +30,14 @@
             }
           ]
         },
+        "outputOptions": {
+          "maximum": 3,
+          "minimum": 0,
+          "type": "integer"
+        },
+        "pseudoCol": {
+          "type": "string"
+        },
         "locationCol": {
           "type": "string"
         },
@@ -106,6 +114,8 @@
       "required": [
         "duplicateDetection",
         "suppressCols",
+        "outputOptions",
+        "pseudoCol",
         "locationCol",
         "datetimeCol",
         "groupByCol",
